Adicionando a primeira versao funcionando do upload e do findrec

+++ app/include/diskManager.h
@@ -29,14 +29,14 @@ public:
     @param id_block representa o ID sequencial do bloco a ser lido
     @param buffer ponteiro para o buffer na RAM onde o bloco vai ser copiado (buffer deve ter tamanho BLOCK_SIZE)
     */
-    void readBlock(int id_block, char* buffer);
+    void readBlock(long long id_block, char* buffer);
 
     /**
     @brief escreve em um bloco de tamanho fixo da RAM pro disco
     @param id_block o ID do bloco que sera escrito/sobreescrito
     @param buffer ponteiro para os dados da RAM que serão escritos no disco
     */
-    void writeBlock(int id_block, const char* buffer);
+    void writeBlock(long long id_block, const char* buffer);
 
     /**
     @brief calcula o numero total de blocos existentes no arquivo
